AUTOLOAD improvements: options and forward declarations
We can now predeclare (forward declaration) subs that will be created
with the AUTOLOAD. Those are all the subs names after the 'AUTOLOAD' in
the import list. This allow to remove parentheses after the sub name.

Subs declared that way also get optional options like the other subs,
but they will be evaluated only when the AUTOLOAD runs for that sub.